backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
asyncpg==0.29.0
psycopg2-binary==2.9.10
redis==5.0.1
celery==5.3.4
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
httpx==0.25.2
python-multipart==0.0.6
python-dotenv==1.0.0
playwright==1.40.0
selenium==4.15.2
fake-useragent==1.4.0
requests==2.31.0